feature 3491: prepare structure for options of indicators
git-svn-id: http://talendforge.org/svn/tis_shared/trunk@13326 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.cwm.management/src/main/java/org/talend/dq/analysis/AnalysisExecutor.java
@@ -68,7 +68,7 @@ public abstract class AnalysisExecutor implements IAnalysisExecutor {
 
         // --- compute execution duration
         long endtime = System.currentTimeMillis();
-        resultMetadata.setExecutionDuration(endtime - startime);
+        resultMetadata.setExecutionDuration((int) (endtime - startime));
         return getReturnCode(ok);
     }
